# starmap-contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: "星图 StarMap API"
  description: "人才能力星云导航系统 - 后端 API 契约"
  version: "1.0.0"
  contact:
    name: "StarMap Team"
  x-summary: "人才技能图谱知识系统全量 API"
servers:
  - url: "http://localhost:8000/api/v1"
    description: "本地开发环境"
  - url: "https://api.starmap.dev/api/v1"
    description: "评审环境"
security:
  - BearerAuth: []
paths:
...
  /positions:
    get:
      operationId: listPositions
      summary: "岗位列表"
      description: "返回系统中所有已发现的岗位"
      tags: [岗位管理]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: industry
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: "岗位列表"
          content:
            application/json:
              schema:
                type: object
                required: [items, total, page, page_size]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/PositionNode"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
  /positions/{position_id}:
    get:
      operationId: getPositionDetail
      summary: "岗位详情"
      description: "获取单个岗位的完整信息，包含技能要求"
      tags: [岗位管理]
      parameters:
        - name: position_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "岗位详情"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionNode"
        "404":
          description: "岗位不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /positions/discover:
    post:
      operationId: discoverPosition
      summary: "触发岗位发现"
      description: "触发自动化岗位发现流水线：涌现检测 → 聚类 → 定义生成 → 幻觉防控"
      tags: [岗位管理]
      responses:
        "200":
          description: "发现完成"
          content:
            application/json:
              schema:
                type: object
                required: [status, emerging_skills, count, skills_analyzed]
                properties:
                  status:
                    type: string
                    example: "completed"
                  emerging_skills:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
                  skills_analyzed:
                    type: integer
        "500":
          description: "发现失败"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ──────────────────────── 图谱查询 ────────────────────────
...
  /graph/ka/{ka_id}/positions:
    get:
      operationId: getKAPositions
      summary: "KA 下的岗位列表"
      description: "返回指定 KnowledgeArea 下的 Position 节点及其与 Skill 的 REQUIRES 关系"
      tags: [图谱查询]
      parameters:
        - name: ka_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "KA 岗位列表"
          content:
            application/json:
              schema:
                type: object
                required: [ka_id, ka_name, positions, position_skill_edges]
                properties:
                  ka_id:
                    type: string
                  ka_name:
                    type: string
                  positions:
                    type: array
                    items:
                      $ref: "#/components/schemas/GraphNode"
                  position_skill_edges:
                    type: array
                    items:
                      $ref: "#/components/schemas/GraphEdge"

  # ──────────────────────── 匹配诊断 ────────────────────────
  /match/position:
    post:
      operationId: matchPosition
      summary: "人岗匹配"
      description: "将个人的技能集与目标岗位进行匹配度计算"
      tags: [匹配诊断]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequest"
      responses:
        "200":
          description: "匹配完成"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
        "400":
          description: "请求参数错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from authentication endpoint"
  schemas:
...
    SkillNode:
      type: object
      required: [skill_id, name, category, proficiency, confidence, source_count]
      properties:
        skill_id:
          type: string
          description: "技能唯一标识（归一化后的 ID）"
          example: "skill_python_3.x"
        name:
          type: string
          description: "技能展示名称"
          example: "Python"
        category:
          type: string
          enum: [hard_skill, soft_skill, tool, certificate]
          description: "技能分类"
          example: "hard_skill"
        proficiency:
          type: string
          enum: ["了解", "熟悉", "精通"]
          description: "熟练度等级"
          example: "精通"
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: "该技能节点置信度"
          example: 0.95
        source_count:
          type: integer
          minimum: 0
          description: "来源文档计数"
          example: 15
    PositionNode:
      type: object
      required: [position_id, name, industry, description, skills_required]
      properties:
        position_id:
          type: string
          example: "pos_backend_engineer"
        name:
          type: string
          example: "后端开发工程师"
        industry:
          type: string
          example: "互联网/IT"
        description:
          type: string
          example: "负责后端服务架构设计与核心模块开发"
        skills_required:
          type: array
          items:
            $ref: "#/components/schemas/SkillNode"
        discovered_at:
          type: string
          format: date-time
          description: "发现时间"
    GraphNode:
      type: object
      required: [id, labels, properties]
      properties:
        id:
          type: string
          example: "skill_python_3.x"
        labels:
          type: array
          items:
            type: string
          example: ["Skill", "HardSkill"]
        properties:
          type: object
          description: "节点属性键值对"
          additionalProperties: true
          example:
            name: "Python"
            category: "hard_skill"
            confidence: 0.95
    GraphEdge:
      type: object
      required: [source_id, target_id, type]
      properties:
        source_id:
          type: string
          example: "skill_python_3.x"
        target_id:
          type: string
          example: "pos_backend_engineer"
        type:
          type: string
          example: "REQUIRED_FOR"
        properties:
          type: object
          description: "边属性"
          additionalProperties: true
          example:
            weight: 0.9
    MatchRequest:
      type: object
      required: [person_skills, target_position]
      properties:
        person_skills:
          type: array
          description: "个人技能列表"
          items:
            type: object
            required: [name]
            properties:
              skill_id:
                type: string
                nullable: true
              name:
                type: string
              category:
                type: string
                default: "hard_skill"
              proficiency:
                type: string
                default: "熟悉"
              confidence:
                type: number
                default: 1.0
              source_count:
                type: integer
                default: 0
        target_position:
          type: string
          description: "目标岗位 ID 或名称"
          example: "pos_backend_engineer"
        options:
          type: object
          properties:
            threshold:
              type: number
              format: float
              minimum: 0
              maximum: 1
              default: 0.6
              description: "匹配阈值，低于此值视为严重不匹配"
    MatchResult:
      type: object
      required: [match_id, target_position, match_score, matched_skills, gap_skills, recommendations]
      properties:
        match_id:
          type: string
        target_position:
          type: string
        match_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: "总体匹配度"
          example: 0.74
        matched_skills:
          type: array
          description: "已匹配技能"
          items:
            type: string
          example: ["Python", "FastAPI", "PostgreSQL"]
        gap_skills:
          type: array
          description: "差距技能（缺失或不足）"
          items:
            type: string
          example: ["Docker", "Kubernetes", "分布式系统设计"]
        recommendations:
          type: array
          description: "学习路径建议"
          items:
            type: string
          example: ["建议优先学习 Docker 容器化部署", "补充分布式系统理论知识"]
        missing_required:
          type: array
          items:
            type: string
        missing_bonus:
          type: array
          items:
            type: string
        skill_gap_detail:
          type: array
          items:
            type: object
            properties:
              skill:
                type: string
              importance:
                type: string
              gap_level:
                type: string
              learning_path:
                type: array
                items:
                  type: string
        overall_assessment:
          type: string
        estimated_learning_time:
          type: string
        cii:
          type: number
          nullable: true
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: "错误描述"
          example: "岗位不存在"
        code:
          type: string
          description: "错误码（可选，部分端点返回）"
          example: "POSITION_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: "错误发生时间（可选）"

    # ──────────────────────── 数据大屏 ────────────────────────
...